requirements.txt
joblib==1.3.1
streamlit==1.25.0
openai==0.27.8
s3fs==2022.11.0
numpy==1.25.0
boto3==1.24.59
requests==2.28.2
faiss-cpu
structlog
streamlit-authenticator
bs4
langchain
